+++ buyer-api/src/facades/index.js
@@ -15,7 +15,7 @@ import closeDataOrderFacade from './closeDataOrderFacade';
 import { getOrdersForBuyer, fetchAndCacheDataOrder } from './getOrdersFacade';
 
 export { getNotaryInfo, getNotariesInfo, fetchAndCacheNotary } from './notariesFacade';
-export { checkAndTransfer } from './transferFacade';
+export { checkAndTransfer, checkInitialRootBuyerFunds, monitorFunds } from './transferFacade';
 
 export {
   createDataOrderFacade,
