+++ buyer-api/src/utils/index.js
@@ -14,4 +14,3 @@ export {
 export { errorHandler, asyncError, validateAddress } from './routes';
 export { wibcoin, dataExchange, DataOrderContract } from './contracts';
 export { delay } from './delay';
-export { checkInitialRootBuyerFunds, monitorFunds } from './funding';
